refactor(api): move route / controller into prescription
